Radix Sort
The Radix sort algorithm works by ordering each digit from least significant to most significant. 
In base 10, radix sort would sort by the digits in the one's place, then the ten's place, and so on.
To sort the values in each digit place, Radix sort employs counting sort as a subroutine.
This means that for a three-digit number in base 10, counting sort will be used to sort the 1st, 10th, and 100th places, resulting in a completely sorted list. Here's a rundown of the counting sort algorithm.